% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_run_destroy_pfilter.R
\name{create_run_destroy_pfilter}
\alias{create_run_destroy_pfilter}
\title{Create, run, and destroy a particle filter}
\usage{
create_run_destroy_pfilter(df, threshold, work_dir, is_forecast, quiet)
}
\arguments{
\item{df}{A data frame holding inputs for an individual cohort.}

\item{threshold}{Threshold for triggering resampling.
A number between 0 and 1.}

\item{work_dir}{The name of the folder where
intermediate calculations are carried out and
results are stored.}

\item{is_forecast}{Whether the particle filter
is being used for a forecast.}

\item{quiet}{Whether to show a message saying calculations
have started.}
}
\value{
Return \code{TRUE} invisibly and
writes the files described above to \code{work_dir}.
}
\description{
Create a particle filter, run it, write the output to disk, then destroy
the particle filter. The particle filter deals with a single
cohort (or, more precisely, a single combination of cohort
and sex/gender.)
}
\details{
\code{create_run_destroy_pfilter} is an internal function,
and end-users should never need to call it directly.
\code{create_run_destroy_pfilter} does not do any
validity-checking on its arguments: error checking
is the responsibility of the calling functions.

An individual particle filter can hold millions of
values, so we avoid creating multiple particle
filters at the same time.

\code{create_run_destroy_pfilter} writes files to directory
\code{work_dir}.

WARNING - THIS DESCRIPTION NEEDS TO BE UPDATED
The first file is called \code{tmp-population-<cohort>-<sexgender>.csv},
where values for \code{<cohort>} and \code{<sexgender>} are
taken from \code{df}. If \code{df} includes a region dimension,
then the file has columns \code{"cohort"}, \code{"sexgender"},
\code{"time"}, \code{"age"}, \code{"region"},
\code{"iteration"}, and \code{"count"}. If \code{df} does not
include a region dimension, then the file does not include a
\code{"region"} column.

Next there is a set of files describing events, each of which
has a filename of the form \code{tmp-<event>-<cohort>-<sexgender>.csv}.
If \code{df} has a region dimension, and if the value for
\code{is_dominant} for \code{df} is \code{TRUE},
then \code{<event>} consists of \code{"births"},
\code{"deaths"}, \code{"internal_in"}, \code{"internal_out"},
\code{"immigration1"}, \code{"emigration1"},
\code{"immigration2"}, and \code{"emigration2"}. If
If \code{df} does not include a region dimension,
then \code{<event>} does not include \code{"internal_in"}
or \code{"internal_out"}. If the value for
\code{is_dominant} for \code{df} is \code{FALSE},
then \code{<event>} does not include \code{"births"}.

If \code{df} includes a region dimension,
then the file has variables \code{"cohort"}, \code{"sexgender"},
\code{"time"}, \code{"age"}, \code{"iteration"},
and \code{"count"}. If \code{df} does not include a region dimension,
then the file does not include variable \code{"region"}.

The final file is called \code{tmp-diagnostic-<cohort>-<sexgender>.csv}.
The file has columns \code{"cohort"}, \code{"sexgender"},
\code{"time"}, \code{"age"}, \code{"triangle"}, \code{"ess"},
\code{"resampled"}, and \code{"n_unique"}.
}
