% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_run_destroy_pfilter.R
\name{do_forecast}
\alias{do_forecast}
\title{Do a forecast for a single cohort}
\usage{
do_forecast(df, work_dir, quiet)
}
\arguments{
\item{df}{A data frame holding inputs for an individual cohort.}

\item{work_dir}{The name of the folder where
intermediate calculations are carried out and
results are stored.}

\item{quiet}{Whether to show a message saying calculations
have started.}
}
\value{
Return \code{TRUE} invisibly and
writes the files described above to \code{work_dir}.
}
\description{
Do a forecast for a single cohort. We draw straight from the
posterior distribution, so we don't need to use particle filtering.
}
\details{
\code{do_forecast} is an internal function,
and end-users should never need to call it directly.
}
