% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpmpf-package.R
\docType{package}
\name{dpmpf-package}
\alias{dpmpf}
\alias{dpmpf-package}
\title{Bayesian demographic account for England and Wales}
\description{
Create a demographic account for England and Wales,
optionally including regions and internal migration.
}
\details{
The package in no longer active development.
}
\author{
\strong{Maintainer}: Office for National Statistics \email{demographic.methods@ons.gov.uk} [copyright holder, funder]

Authors:
\itemize{
  \item John Bryant
  \item Nick Grant \email{h@ckte.ch}
}

Other contributors:
\itemize{
  \item Aidan Metcalfe [contributor]
}

}
\keyword{internal}
