% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdm-constructors.R
\name{new_CdmNoregPoibin}
\alias{new_CdmNoregPoibin}
\title{Create an object of class Rcpp_CdmNoregPoibin}
\usage{
new_CdmNoregPoibin(counts_data, prob)
}
\arguments{
\item{counts_data}{A numeric vector of non-negative whole numbers.
\code{NA}s are allowed.}

\item{prob}{A numeric scalar between 0 and 1 (exclusive).}
}
\value{
An S4 object with class "Rcpp_CdmNoregPoibin"
}
\description{
Create an object of (C++) class \code{Rcpp_CdmNoregPoibin}.
Objects of class \code{Rcpp_CdmNoregPoibin} are cohort-level
data models, plus cohort-level data, for a single dataset. The data
does not have a region dimension. The data model
is based on a Poisson-binomial mixture.
}
\details{
\code{new_CdmNoregPoibin} is an internal function, and should
never be needed by end users.
}
