% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdms-constructors.R
\name{new_CdmsNoreg}
\alias{new_CdmsNoreg}
\title{Create an object of class Rcpp_CdmsNoreg}
\usage{
new_CdmsNoreg(children = list())
}
\arguments{
\item{children}{A list of objects of class "CdmNoreg"}
}
\value{
An object of class \code{Rcpp_CdmsNoreg}
}
\description{
Objects of class \code{Rcpp_CdmsNoreg} hold
cohort no-region data models for a single demographic
series, eg population or deaths.
}
\seealso{
\code{\link{new_CdmNoregPoibin}},
\code{\link{new_CdmNoregNbinom}},
\code{\link{new_CdmsWithreg}}
}
