% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdms-constructors.R
\name{new_CdmsWithreg}
\alias{new_CdmsWithreg}
\title{Create an object of class Rcpp_CdmsNoreg}
\usage{
new_CdmsWithreg(children = list())
}
\arguments{
\item{children}{A list of objects of class "CdmWithreg"}
}
\value{
An object of class \code{Rcpp_CdmsWithreg}
}
\description{
Objects of class \code{Rcpp_CdmsNoreg} hold
cohort with-region data models for a single demographic
series, eg population or deaths.
}
\seealso{
\code{\link{new_CdmWithregPoibin}},
\code{\link{new_CdmWithregNbinom}},
\code{\link{new_CdmsNoreg}}
}
