% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdm-constructors.R
\name{new_CdmWithregNbinom}
\alias{new_CdmWithregNbinom}
\title{Create an object of class Rcpp_CdmWithregNbinom}
\usage{
new_CdmWithregNbinom(counts_data, ratio, disp)
}
\arguments{
\item{counts_data}{A numeric matrix of non-negative whole numbers.
\code{NA}s are allowed.}

\item{ratio}{A numeric matrix of non-negative numbers, with the
same dimensions as \code{counts_data}. \code{NA}s are allowed
at positions where \code{counts_data} is \code{NA}.}

\item{disp}{A numeric matrix of non-negative numbers, with the
same dimensions as \code{counts_data}. \code{NA}s are allowed
at positions where \code{counts_data} is \code{NA}.}
}
\value{
An S4 object with class "Rcpp_CdmWithregNbinom"
}
\description{
Create an object of (C++) class \code{Rcpp_CdmWithregNbinom}.
Objects of class \code{Rcpp_CdmWithregNbinom} are cohort-level
data models, plus cohort-level data, for a single dataset. The data
has a region dimension. The data model
is based on a negative binomial distribution.
}
\details{
\code{new_CdmWithregNbinom} is an internal function, and should
never be needed by end users.
}
