% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdm-constructors.R
\name{new_CdmWithregPoibin}
\alias{new_CdmWithregPoibin}
\title{Create an object of class Rcpp_CdmWithregPoibin}
\usage{
new_CdmWithregPoibin(counts_data, prob)
}
\arguments{
\item{counts_data}{A numeric matrix of non-negative whole numbers.
\code{NA}s are allowed.}

\item{prob}{A numeric scalar between 0 and 1 (exclusive).}
}
\value{
An S4 object with class "Rcpp_CdmWithregPoibin"
}
\description{
Create an object of (C++) class \code{Rcpp_CdmWithregPoibin}.
Objects of class \code{Rcpp_CdmWithregPoibin} are cohort-level
data models, plus cohort-level data, for a single dataset. The data
has a region dimension. The data model
is based on a Poisson-binomial mixture.
}
\details{
\code{new_CdmWithregPoibin} is an internal function, and should
never be needed by end users.
}
